requirements.txt
streamlit
spacy
PyMuPDF
scikit-learn
python-dotenv
pandas
numpy
langchain
langchain-google-genai
google-generativeai
pydantic
